Fix tier_active() function
I got the torrent state wrong, it was called pausedUP. A secondary yet major error is that when the tier increases it sets a new ratio_limit and the function was never fulfilling the condition ratio >= ratio_limit, instead it had to read the old ratio_limit and compare itself with that one. Now it also checks if the tier is between 1 and 9, because the torrent will not find itself paused on tier 0 without manual intervention.