Add credentials support
Now the script can read credentials from the configuration file and no longer needs to hard code them. The credentials are no longer given directly to qbittorrentapi.Client(), instead auth_log_in() is used, this way the default python trace back error is replaced with a custom message that explains where the process went wrong. To make the code safer when authentication bypass is on the script calls auth_log_in() with empty parameters, because otherwise is_logged_in() fails to check the log in status.